Fixig Variable's Names

+++ Card.java
@@ -1,7 +1,7 @@
 class Card extends Payment{
     Integer cvv;
 
-    public Card (Integer cvv, String name, Double accNum) {
+    public Card (Integer cvv, String name, Integer accNum) {
         super(name, accNum);
         this.cvv = cvv;
     }
