+++ Paypal.java
@@ -1,7 +1,7 @@
 class Paypal extends Payment {
     String correo;
 
-    public Paypal (String name, Double accNum, String correo) {
+    public Paypal (String name, Integer accNum, String correo) {
         super(name, accNum);
         this.correo = correo;
     }
